Add API for Picture plane data access as slice
The `plane_data()` method is replaced by a `plane()` method returning a `Plane`
which can be dereferenced as a slice. This is more convenient when safely
copying the decoded picture.

The plane data pointer is now accessible through the `plane_data_ptr()` method.

With contributions from Sebastian Dröge <sebastian@centricular.com>.