Improve API for properly handling errors and unreffing data when needed
On errors it is necessary to unref the data passed in, except for EAGAIN
where the remaining data is still queued up in the decoder and must be
kept alive and has to be continued with at a later time after some
decoded pictures are handled.

The existing decode() API did not allow handling this correctly, and
especially not recovering from errors from getting a decoded picture
while data was still queued up.