Fix struct padding size depending on features
fb0355b421609484979ae877d4a1cf2f62db626c introduced the v1_1 feature,
but the size of the reserved field in Dav1dSettings was mixed up.  It
isn’t much of an issue when using this new feature as only four bytes
are wasted, but it very much is when not using it, as four bytes of
garbage are passed to the functions taking it as a pointer.

I also added a test to avoid reproducing this mistake in the future.